# docs/openapi/api-v2.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lurus API V2 - Multi-Tenant Gateway",
...
    "version": "2.0.0",
    "contact": {
      "name": "Lurus Support",
      "email": "support@lurus.cn"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://lurus.cn/terms"
    }
  },
  "servers": [
    {
      "url": "https://hub.lurus.cn",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v2/{tenant_slug}/logs": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "Get User Logs",
        "description": "Gets usage logs for the current user",
        "operationId": "getLogsV2",
        "security": [
          {
            "OIDCBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantSlug"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "model",
            "in": "query",
            "description": "Filter by model name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "description": "Start time (Unix timestamp)",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "description": "End time (Unix timestamp)",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v2/{tenant_slug}/logs/all": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "Get All Logs",
        "description": "Gets all users' logs (Admin only)",
        "operationId": "getAllLogsV2",
        "security": [
          {
            "OIDCBearer": [
              "admin"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantSlug"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "Filter by user ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "OIDCBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token issued by the deployment's configured OIDC provider. Which provider that is, is deploy-time configuration; no specific vendor is part of this contract."
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie for platform admin"
      }
    },
    "parameters": {
      "TenantSlug": {
        "name": "tenant_slug",
        "in": "path",
        "required": true,
        "description": "Tenant identifier (slug). On routes behind TenantSlugGuard the reserved value `~` means \"the tenant this session authenticated into\"; the server resolves it from the session, and it is what the console sends. An explicit slug must equal that tenant (403 TENANT_MISMATCH otherwise). `GET /user/me` reports the real slug as `tenant_slug`.",
        "schema": {
          "type": "string",
          "pattern": "^([a-z0-9-]+|~)$",
          "example": "lurus"
        }
      },
      "Page": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-based)",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "PageSize": {
        "name": "page_size",
        "in": "query",
        "description": "Items per page",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Invalid request"
          }
        }
      },
...
      "LogList": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "user_id": {
                  "type": "integer"
                },
                "model": {
                  "type": "string"
                },
                "prompt_tokens": {
                  "type": "integer"
                },
                "completion_tokens": {
                  "type": "integer"
                },
                "quota": {
                  "type": "integer"
                },
                "created_at": {
                  "type": "integer"
                }
              }
            }
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}